Temporary solution for play functionality
HTML rebuilt for two players, but the JS is not. Each players element
has its own ID, so as a temporary fix to still see if the game works, I
added "player1" to each getElementByID that needed it. This must be
resolved in later updates of the JS.

(Also, due to IDEs having differing setups, single and double qoutes are
auto formatted differently, as is trailing commas for example.)